Fix path-view layout by using display:flex
The previous change to make the path-view table scrollable was not
working correctly because the javascript was setting the display
property to "block" instead of "flex". This commit fixes the
javascript to set the display property to "flex" and cleans up the
inline styles in index.html.

Gemini commit: The edit button is still below the path table div on my display